Pond Art and Prefabs
I went through and added the separate pieces of art for the pond into folders, and I created an environment folder in the prefabs folder where I put the prefabs for the pond level. The prefabs that I created all should have colliders on them as well. I didn't make any scene changes.